Add DARTH workgroup citation to the template
I copy pasted this structure for a project with a student but I had to add a citation to all the files. I would be useful to have the DARTH citation already in the template to users do not have to cite DARTH themselves. I added a suggested way of citing.